src/main/webapp/app/entities/accessSampleApplication/table-3/delete/table-3-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { Table3Service } from '../service/table-3.service';

import { Table3DeleteDialogComponent } from './table-3-delete-dialog.component';

describe('Table3 Management Delete Component', () => {
  let comp: Table3DeleteDialogComponent;
  let fixture: ComponentFixture<Table3DeleteDialogComponent>;
  let service: Table3Service;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, Table3DeleteDialogComponent],
      providers: [NgbActiveModal],
    })
      .overrideTemplate(Table3DeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(Table3DeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(Table3Service);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete(123);
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith(123);
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
